notepal/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
black==23.1.0
boto3==1.26.70
botocore==1.29.70
Brotli==1.1.0
bson==0.5.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.3
cssselect2==0.7.0
dataclasses-json==0.5.13
Django==4.1.7
django-appconf==1.0.5
django-browser-reload==1.11.0
django-contrib-postgres==0.0.1
django-cryptography==1.1
django-ninja==0.22.2
django-tailwind==3.6.0
docopt==0.6.2
exceptiongroup==1.1.2
Faker==18.4.0
filelock==3.12.2
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
langchain==0.0.238
langsmith==0.0.12
lxml==4.9.3
Markdown==3.5
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
md2pdf==1.0.1
mistletoe==1.2.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
nvidia-pyindex==1.0.9
openai==0.27.8
openapi-schema-pydantic==1.2.4
oscrypto==1.3.0
packaging==23.0
pathspec==0.11.0
pgvector==0.1.8
Pillow==10.0.0
platformdirs==3.0.0
pluggy==1.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.12.2
PyPDF4==1.27.0
pyphen==0.14.0
pypng==0.20220715.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-pdb==0.3.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.6.3
reportlab==3.6.13
requests==2.28.2
s3transfer==0.6.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.3
svglib==1.5.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing-extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.1
uritools==4.0.2
urllib3==1.26.14
weasyprint==60.1
webencodings==0.5.1
xhtml2pdf==0.2.11
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.17.0
zopfli==0.2.3
